move files around for cleaner paths
