+++ frontend/src/components/ui/config-details.ts
@@ -24,7 +24,7 @@ import { isApiError } from "@/utils/api";
 import { DEPTH_SUPPORTED_SCOPES, isPageScopeType } from "@/utils/crawler";
 import { humanizeSchedule } from "@/utils/cron";
 import { pluralOf } from "@/utils/pluralize";
-import { richText } from "@/utils/rich-text/rich-text";
+import { richText } from "@/utils/rich-text";
 import { getServerDefaults } from "@/utils/workflow";
 
 /**
